requirements.txt
langchain==0.3.13
langchain_groq==0.2.2
python-dotenv==1.0.1